refactor: split serve into wake, server, and scheduler packages
Move the HTTP server, the cron scheduler, and the shared wake primitive
out of the cmd package into packages named by capability:

- wake: Waker type, Broadcast, and ByName (machine name to MAC)
- server: web interface (handlers, pinger, templates)
- scheduler: cron wake-ups

The cmd package is now thin command wiring. Build version vars stay in
cmd so the release ldflags path is unchanged, and are passed to the
server as a BuildInfo value. The send command now reuses wake.Broadcast
instead of building the magic packet itself.

No behavior change.
