+++ scheduler/scheduler.go
@@ -1,4 +1,6 @@
-package cmd
+// Package scheduler runs cron-driven wake-ups for configured machines while the
+// serve command is running.
+package scheduler
 
 import (
 	"fmt"
@@ -7,6 +9,7 @@ import (
 
 	"github.com/robfig/cron/v3"
 	"github.com/trugamr/wol/config"
+	"github.com/trugamr/wol/wake"
 )
 
 // resolvedSchedule is a config.Schedule that has been validated against the
@@ -31,7 +34,7 @@ func resolveSchedules(machines []config.Machine, schedules []config.Schedule) ([
 			label = s.Machine
 		}
 
-		mac, err := getMacByName(machines, s.Machine)
+		mac, err := wake.ByName(machines, s.Machine)
 		if err != nil {
 			return nil, fmt.Errorf("schedule %q: %w", label, err)
 		}
@@ -49,24 +52,24 @@ func resolveSchedules(machines []config.Machine, schedules []config.Schedule) ([
 // wakeJob returns the function cron runs when a schedule fires: it logs the
 // trigger and wakes the machine, logging any error. It is split out so the job
 // body can be tested without waiting on cron timing.
-func wakeJob(wake waker, label string, mac net.HardwareAddr) func() {
+func wakeJob(waker wake.Waker, label string, mac net.HardwareAddr) func() {
 	return func() {
 		log.Printf("Triggered scheduled wake for %s", label)
-		if err := wake(mac); err != nil {
+		if err := waker(mac); err != nil {
 			log.Printf("Error sending scheduled magic packet for %s: %v", label, err)
 		}
 	}
 }
 
-// scheduler runs cron wake-ups for the duration of the serve command.
-type scheduler struct {
+// Scheduler runs cron wake-ups for the duration of the serve command.
+type Scheduler struct {
 	cron *cron.Cron
 }
 
-// newScheduler resolves and registers every schedule, returning an error if any
+// New resolves and registers every schedule, returning an error if any
 // references an unknown machine or has an invalid cron expression. The returned
-// scheduler is not started yet; call start to begin firing jobs.
-func newScheduler(machines []config.Machine, schedules []config.Schedule, wake waker) (*scheduler, error) {
+// Scheduler is not started yet; call Start to begin firing jobs.
+func New(machines []config.Machine, schedules []config.Schedule, waker wake.Waker) (*Scheduler, error) {
 	resolved, err := resolveSchedules(machines, schedules)
 	if err != nil {
 		return nil, err
@@ -76,17 +79,17 @@ func newScheduler(machines []config.Machine, schedules []config.Schedule, wake w
 	for _, rs := range resolved {
 		// rs.schedule is already parsed, so register it directly rather than
 		// re-parsing the spec string via AddFunc.
-		c.Schedule(rs.schedule, cron.FuncJob(wakeJob(wake, rs.label, rs.mac)))
+		c.Schedule(rs.schedule, cron.FuncJob(wakeJob(waker, rs.label, rs.mac)))
 	}
 
-	return &scheduler{cron: c}, nil
+	return &Scheduler{cron: c}, nil
 }
 
-// start begins running the scheduled jobs in the background.
+// Start begins running the scheduled jobs in the background.
 //
-// There is intentionally no stop method: serve blocks on http.ListenAndServe
-// for the life of the process, so the scheduler simply dies with it. Add a stop
+// There is intentionally no Stop method: serve blocks on http.ListenAndServe
+// for the life of the process, so the scheduler simply dies with it. Add a Stop
 // method (wrapping s.cron.Stop) when serve grows graceful shutdown.
-func (s *scheduler) start() {
+func (s *Scheduler) Start() {
 	s.cron.Start()
 }
